Fix a few warnings

+++ src/util/filterdb.cpp
@@ -11,7 +11,7 @@
 #include <omp.h>
 #endif
 
-int dofilter(std::string inputDb, std::string outputDb, int threads, int column, std::string regexStr) {
+int dofilter(std::string inputDb, std::string outputDb, int threads, size_t column, std::string regexStr) {
     DBReader<std::string>* dataDb=new DBReader<std::string>(inputDb.c_str(),(std::string(inputDb).append(".index")).c_str());
     dataDb->open(DBReader<std::string>::LINEAR_ACCCESS);
     DBWriter* dbw = new DBWriter(outputDb.c_str(), (std::string(outputDb).append(".index")).c_str(), threads);
@@ -101,6 +101,6 @@ int filterdb(int argn, const char **argv)
     return dofilter(par.db1,
                     par.db2,
                     par.threads,
-                    par.filterColumn,
+                    static_cast<size_t>(par.filterColumn),
                     par.filterColumnRegex);
 }
